move examples outside src
